requirements.txt
discord.py
pandas
requests
stockstats
python-dotenv
apscheduler